Refactor compta package into module subdirectories
