+++ migrations.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from MOTEUR.compta.achat_db import init_db as init_purchase
+from MOTEUR.compta.achats.db import init_db as init_purchase
 
 
 def apply_migrations(db_path: Path | str) -> None:
